requirements.txt
streamlit
pandas
pillow[pil]
qrcode
